Correct type of size parameter.

+++ sysdeps/generic/shmget.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 1995, 1996, 1997 Free Software Foundation, Inc.
+/* Copyright (C) 1995, 1996, 1997, 1999 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
    Contributed by Ulrich Drepper <drepper@gnu.ai.mit.edu>, August 1995.
 
@@ -26,7 +26,7 @@
 int
 shmget (key, size, shmflg)
      key_t key;
-     int size;
+     size_t size;
      int shmflg;
 {
   __set_errno (ENOSYS);
